A few Major Changes
1. Location Data
-Data is now obtained using the Google Maps Object. This allows for a
much clearer implementation as well as not requiring us to constantly be
requesting updates from the phone.

2. Spinner Item Selection
-Added a default item on the list that does not do anything.
-Added some functionality to the PushPin option

3. Upcoming Features
PushPin will display an alert box to allow the user to enter text.
<Possible Limit on Character Usage>
About us and Instructions will lead to a different page on the phone
with clear examples. More research is necessary on how to implement
this.